# packages/api-client/src/generated/models/settingsUpdateDto.ts
/**
 * Generated by orval v8.19.0 🍺
 * Do not edit manually.
 * Agentic Dev Kit API
 * Auth, tasks, and AI (copilot/speech) endpoints.
 * OpenAPI spec version: 0.1.0
 */
import type { SettingsUpdateDtoAi } from './settingsUpdateDtoAi';
import type { SettingsUpdateDtoGeneral } from './settingsUpdateDtoGeneral';
import type { SettingsUpdateDtoSecrets } from './settingsUpdateDtoSecrets';

export interface SettingsUpdateDto {
  ai?: SettingsUpdateDtoAi;
  general?: SettingsUpdateDtoGeneral;
  secrets?: SettingsUpdateDtoSecrets;
}